🔨 [Refactor] 버튼, 텍스트필드 뷰 독립
